+++ src/base/lib/Convex/NodeQueries.hs
@@ -7,10 +7,13 @@ module Convex.NodeQueries(
   queryEraHistory,
   querySystemStart,
   queryLocalState,
-  queryTip
+  queryTip,
+  queryProtocolParameters
 ) where
 
-import           Cardano.Api                                        (CardanoMode,
+import           Cardano.Api                                        (BabbageEra,
+                                                                     BundledProtocolParameters,
+                                                                     CardanoMode,
                                                                      ChainPoint,
                                                                      ConsensusModeParams (..),
                                                                      Env (..),
@@ -91,3 +94,15 @@ queryLocalState query connectInfo = do
     Left err -> do
       fail ("queryLocalState: Failed with " <> show err)
     Right result -> pure result
+
+queryProtocolParameters :: LocalNodeConnectInfo CardanoMode -> IO (BundledProtocolParameters BabbageEra)
+queryProtocolParameters connectInfo = do
+  result <- queryLocalState (CAPI.QueryInEra CAPI.BabbageEraInCardanoMode (CAPI.QueryInShelleyBasedEra CAPI.ShelleyBasedEraBabbage CAPI.QueryProtocolParameters)) connectInfo
+  case result of
+    Left err -> do
+      fail ("queryProtocolParameters: failed with: " <> show err)
+    -- Right k -> pure (CAPI.bundleProtocolParams k)
+    Right x -> case CAPI.bundleProtocolParams CAPI.BabbageEra x of
+      Left err -> do
+        fail ("queryProtocolParameters: bundleProtocolParams failed with: " <> show err)
+      Right k -> pure k