+++ changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - More fine-grained functions in `Convex.Wallet.Operator` and `Convex.Query` signing & balancing transactions
 - JSON instances for `CoinSelectionError` and `BalancingError`
+- Add `Convex.Devnet.CardanoNode.withCardanoNodeDevnetConfig` and associated types for modifications to the genesis config files. Includes `allowLargeTransactions` for large transactions.
 
 ### Deleted